do lambda things better
